satops_procedures/requirements.txt
Django>=4.2,<5
PyYAML>=6.0
gunicorn>=21.0
